+++ src/assets/sprites/atlas/lvl4mob-atlas.json
@@ -70,7 +70,7 @@
       "frame": {
           "w": 48,
           "h": 48,
-          "x": 0,
+          "x": 144,
           "y": 48
       },
       "anchor": {
